nextjs-flask/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
langchain-openai==0.2.9
langchain==0.3.7